Fix mypy type error in count_matches function
Changed return type annotation from int to float to match the actual
return value from count_pattern_matches which returns a weighted_count
as float.

This fixes the mypy error:
app/repair_pipeline.py:403: error: Incompatible return value type
(got "float", expected "int")

+++ ml_service/app/repair_pipeline.py
@@ -397,7 +397,7 @@ for context_type, templates in CONTEXT_TEMPLATES.items():
 print("Модель готова к использованию.\n")
 
 
-def count_matches(patterns: List, text: str) -> int:
+def count_matches(patterns: List, text: str) -> float:
     """Count pattern matches in text (wrapper for tests)."""
     count, _ = count_pattern_matches(patterns, text)
     return count
